refactor: 테스트 코드 패키지 및 클래스명 변경
- 'app' 패키지의 테스트 코드를 'service' 패키지로 이동
- 'AppApplicationTests' 클래스명을 'ApplicationTests'로 수정
- 'SentryIntegrationTest'의 예외 메시지에서 'app-api' 제거
