+++ service-api/src/test/kotlin/com/millo/ollim/service/SentryIntegrationTest.kt
@@ -1,4 +1,4 @@
-package com.millo.ollim.app
+package com.millo.ollim.service
 
 import io.sentry.Sentry
 import org.junit.jupiter.api.Test
@@ -10,7 +10,7 @@ class SentryIntegrationTest {
     @Test
     fun `Sentry에 예외 전송 테스트`() {
         try {
-            throw Exception("Sentry app-api 테스트 예외입니다.")
+            throw Exception("Sentry 테스트 예외입니다.")
         } catch (e: Exception) {
             Sentry.captureException(e)
         }
